+++ android/src/main/kotlin/io/github/nullptrx/pangleflutter/PangleAdSlotManager.kt
@@ -13,8 +13,6 @@ object PangleAdSlotManager {
     slotId: String,
     imgSize: TTSize,
     isSupportDeepLink: Boolean,
-    splashButtonType: Int,
-    downloadType: Int
   ): AdSlot {
 
     return AdSlot.Builder().apply {
@@ -22,10 +20,7 @@ object PangleAdSlotManager {
       setSupportDeepLink(isSupportDeepLink)
       imgSize.apply {
         setImageAcceptedSize(width.dp, height.dp)
-      } // setSplashButtonType(TTAdConstant.SPLASH_BUTTON_TYPE_FULL_SCREEN)
-      // setDownloadType(TTAdConstant.DOWNLOAD_TYPE_NO_POPUP)
-      setSplashButtonType(splashButtonType)
-      setDownloadType(downloadType)
+      }
     }.build()
   }
 
@@ -36,7 +31,6 @@ object PangleAdSlotManager {
     isVertical: Boolean,
     isSupportDeepLink: Boolean,
     extra: String?,
-    downloadType: Int
   ): AdSlot {
 
     return AdSlot.Builder().apply { // 必选参数 设置您的CodeId
@@ -53,13 +47,12 @@ object PangleAdSlotManager {
       extra?.also {
         setMediaExtra(it)
       }
-      setDownloadType(downloadType)
     }.build()
   }
 
 
   fun getBannerAdSlot(
-    slotId: String, expressSize: TTSizeF, count: Int, isSupportDeepLink: Boolean, downloadType: Int
+    slotId: String, expressSize: TTSizeF, count: Int, isSupportDeepLink: Boolean
   ): AdSlot {
 
     return AdSlot.Builder().apply {
@@ -68,12 +61,11 @@ object PangleAdSlotManager {
       setExpressViewAcceptedSize(expressSize.width, expressSize.height)
       setAdCount(count)
       setSupportDeepLink(isSupportDeepLink)
-      setDownloadType(downloadType)
     }.build()
   }
 
   fun getFeedAdSlot(
-    slotId: String, expressSize: TTSizeF, count: Int, isSupportDeepLink: Boolean, downloadType: Int
+    slotId: String, expressSize: TTSizeF, count: Int, isSupportDeepLink: Boolean
   ): AdSlot {
     return AdSlot.Builder().apply {
       setCodeId(slotId)
@@ -81,12 +73,11 @@ object PangleAdSlotManager {
       // setNativeAdType()
       setAdCount(count)
       setExpressViewAcceptedSize(expressSize.width, expressSize.height)
-      setDownloadType(downloadType)
     }.build()
   }
 
   fun getInterstitialAdSlot(
-    slotId: String, expressSize: TTSizeF, isSupportDeepLink: Boolean, downloadType: Int
+    slotId: String, expressSize: TTSizeF, isSupportDeepLink: Boolean
   ): AdSlot {
 
     return AdSlot.Builder().apply {
@@ -94,7 +85,6 @@ object PangleAdSlotManager {
       setExpressViewAcceptedSize(expressSize.width, expressSize.height)
       setSupportDeepLink(isSupportDeepLink) //请求原生广告时候，请务必调用该方法，设置参数为TYPE_BANNER或TYPE_INTERACTION_AD
       setAdCount(1)
-      setDownloadType(downloadType)
     }.build()
   }
 
@@ -103,8 +93,7 @@ object PangleAdSlotManager {
     slotId: String,
     expressSize: TTSizeF,
     orientation: PangleOrientation,
-    isSupportDeepLink: Boolean,
-    downloadType: Int
+    isSupportDeepLink: Boolean
   ): AdSlot {
 
     return AdSlot.Builder().apply { // 必选参数 设置您的CodeId
@@ -114,12 +103,11 @@ object PangleAdSlotManager {
       setSupportDeepLink(isSupportDeepLink) //设置期望视频播放的方向，为TTAdConstant.HORIZONTAL或TTAdConstant.VERTICAL
       //      setOrientation(if (isVertical) TTAdConstant.VERTICAL else TTAdConstant.HORIZONTAL)
       setOrientation(orientation.ordinal) //激励视频奖励透传参数，字符串，如果用json对象，必须使用序列化为String类型,可为空
-      setDownloadType(downloadType)
     }.build()
   }
 
   fun getNativeBannerAdSlot(
-    slotId: String, size: TTSize, count: Int, isSupportDeepLink: Boolean, downloadType: Int
+    slotId: String, size: TTSize, count: Int, isSupportDeepLink: Boolean
   ): AdSlot {
 
     return AdSlot.Builder().apply {
@@ -127,7 +115,6 @@ object PangleAdSlotManager {
       setImageAcceptedSize(size.width, size.height)
       setAdCount(count)
       setSupportDeepLink(isSupportDeepLink)
-      setDownloadType(downloadType)
     }.build()
   }
 }