+++ android/src/main/kotlin/io/github/nullptrx/pangleflutter/PangleFlutterPlugin.kt
@@ -174,11 +174,8 @@ open class PangleFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
         val hideSkipButton = call.argument<Boolean>("hideSkipButton")
         val isSupportDeepLink = call.argument<Boolean>("isSupportDeepLink") ?: true
         val imgSize = TTSize(1080, 1920)
-        val splashButtonType =
-          call.argument<Int>("splashButtonType") ?: TTAdConstant.SPLASH_BUTTON_TYPE_FULL_SCREEN
-        val downloadType = call.argument<Int>("downloadType") ?: TTAdConstant.DOWNLOAD_TYPE_NO_POPUP
         val adSlot = PangleAdSlotManager.getSplashAdSlot(
-          slotId, imgSize, isSupportDeepLink, splashButtonType, downloadType
+          slotId, imgSize, isSupportDeepLink
         )
         pangle.loadSplashAd(adSlot, FLTSplashAd(hideSkipButton, activity) {
           result.success(it)
@@ -216,9 +213,8 @@ open class PangleFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
         val w: Float = expressArgs.getValue("width").toFloat()
         val h: Float = expressArgs.getValue("height").toFloat()
         val expressSize = TTSizeF(w, h)
-        val downloadType = call.argument<Int>("downloadType") ?: TTAdConstant.DOWNLOAD_TYPE_NO_POPUP
         val adSlot = PangleAdSlotManager.getBannerAdSlot(
-          slotId, expressSize, count, isSupportDeepLink, downloadType
+          slotId, expressSize, count, isSupportDeepLink
         )
         pangle.loadBanner2ExpressAd(adSlot) {
           result.success(it)
@@ -233,9 +229,8 @@ open class PangleFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
         val w: Float = expressArgs.getValue("width").toFloat()
         val h: Float = expressArgs.getValue("height").toFloat()
         val expressSize = TTSizeF(w, h)
-        val downloadType = call.argument<Int>("downloadType") ?: TTAdConstant.DOWNLOAD_TYPE_NO_POPUP
         val adSlot = PangleAdSlotManager.getFeedAdSlot(
-          slotId, expressSize, count, isSupportDeepLink, downloadType
+          slotId, expressSize, count, isSupportDeepLink
         )
         pangle.loadFeedExpressAd(adSlot) {
           result.success(it)
@@ -261,10 +256,9 @@ open class PangleFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
         val w: Float = expressArgs.getValue("width").toFloat()
         val h: Float = expressArgs.getValue("height").toFloat()
         val expressSize = TTSizeF(w, h)
-        val downloadType = call.argument<Int>("downloadType") ?: TTAdConstant.DOWNLOAD_TYPE_NO_POPUP
-        
+
         val adSlot = PangleAdSlotManager.getInterstitialAdSlot(
-          slotId, expressSize, isSupportDeepLink, downloadType
+          slotId, expressSize, isSupportDeepLink
         )
         pangle.loadInteractionExpressAd(adSlot, FLTInterstitialExpressAd(activity) {
           result.success(it)
@@ -323,7 +317,6 @@ open class PangleFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
     val w: Float = expressArgs.getValue("width").toFloat()
     val h: Float = expressArgs.getValue("height").toFloat()
     val expressSize = TTSizeF(w, h)
-    val downloadType = call.argument<Int>("downloadType") ?: TTAdConstant.DOWNLOAD_TYPE_NO_POPUP
     val adSlot = PangleAdSlotManager.getRewardVideoAdSlot(
       slotId,
       expressSize,
@@ -331,7 +324,6 @@ open class PangleFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
       isVertical,
       isSupportDeepLink,
       extra,
-      downloadType
     )
     
     PangleAdManager.shared.loadRewardVideoAd(adSlot, activity, loadingType) {
@@ -352,9 +344,8 @@ open class PangleFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
     val w: Float = expressArgs.getValue("width").toFloat()
     val h: Float = expressArgs.getValue("height").toFloat()
     val expressSize = TTSizeF(w, h)
-    val downloadType = call.argument<Int>("downloadType") ?: TTAdConstant.DOWNLOAD_TYPE_NO_POPUP
     val adSlot = PangleAdSlotManager.getFullScreenVideoAdSlot(
-      slotId, expressSize, orientation, isSupportDeepLink, downloadType
+      slotId, expressSize, orientation, isSupportDeepLink
     )
     
     PangleAdManager.shared.loadFullscreenVideoAd(adSlot, activity, loadingType) {