+++ android/src/main/kotlin/io/github/nullptrx/pangleflutter/view/FlutterSplashView.kt
@@ -38,9 +38,6 @@ class FlutterSplashView(
       val isSupportDeepLink = params["isSupportDeepLink"] as? Boolean ?: true
       val tolerateTimeout = params["tolerateTimeout"] as Double?
       hideSkipButton = params["hideSkipButton"] as? Boolean ?: false
-      val splashButtonType =
-        params["splashButtonType"] as Int? ?: TTAdConstant.SPLASH_BUTTON_TYPE_FULL_SCREEN
-      val downloadType = params["downloadType"] as Int? ?: TTAdConstant.DOWNLOAD_TYPE_NO_POPUP
 
       val imgArgs: Map<String, Int?> = params["imageSize"]?.asMap() ?: mapOf()
       val w: Int = imgArgs["width"] ?: 1080
@@ -50,8 +47,6 @@ class FlutterSplashView(
         slotId,
         imgSize,
         isSupportDeepLink,
-        splashButtonType,
-        downloadType
       )
       PangleAdManager.shared.loadSplashAd(adSlot, this, timeout = tolerateTimeout)
     }