+++ example/lib/page/cn/express/feed_express_page.dart
@@ -168,7 +168,6 @@ class _FeedExpressPageState extends State<FeedExpressPage> {
       android: AndroidFeedConfig(
         slotId: kFeedExpressId375x284,
         expressSize: expressSize,
-        downloadType: PangleDownloadType.none,
         // slotId: kFeedId,
       ),
     );