+++ lib/src/config_android.dart
@@ -112,8 +112,6 @@ class AndroidSplashConfig implements Config {
   final bool isExpress;
   final bool isSupportDeepLink;
   final PangleExpressSize? expressSize;
-  final PangleSplashButtonType splashButtonType;
-  final PangleDownloadType downloadType;
 
   /// The splash ad config for Android
   ///
@@ -123,8 +121,6 @@ class AndroidSplashConfig implements Config {
   /// [isSupportDeepLink] optional. Whether to support deeplink. Default true.
   /// [isExpress] 开屏广告无模板渲染，默认false
   /// [expressSize] optional. 模板宽高
-  /// [splashButtonType] optional, 可点击广告区域，默认全屏
-  /// [downloadType] optional, 下载类广告是否展示二次确认弹窗
   const AndroidSplashConfig({
     required this.slotId,
     this.tolerateTimeout,
@@ -132,8 +128,6 @@ class AndroidSplashConfig implements Config {
     this.isSupportDeepLink = true,
     this.isExpress = false,
     this.expressSize,
-    this.splashButtonType = PangleSplashButtonType.fullScreen,
-    this.downloadType = PangleDownloadType.none,
   });
 
   /// Convert config to json
@@ -146,8 +140,6 @@ class AndroidSplashConfig implements Config {
       'isSupportDeepLink': isSupportDeepLink,
       'isExpress': isExpress,
       'expressSize': expressSize?.toJson(),
-      'splashButtonType': splashButtonType.index,
-      'downloadType': downloadType.index,
     };
   }
 }
@@ -160,7 +152,6 @@ class AndroidRewardedVideoConfig implements Config {
   final bool isSupportDeepLink;
   final PangleLoadingType? loadingType;
   final PangleExpressSize? expressSize;
-  final PangleDownloadType downloadType;
 
   /// The rewarded video ad config for Android
   ///
@@ -176,7 +167,6 @@ class AndroidRewardedVideoConfig implements Config {
   /// [isSupportDeepLink] optional. Whether to support deeplink. default true.
   /// [loadingType] optional. 加载广告的类型，默认[PangleLoadingType.normal]
   /// [expressSize] optional. 模板宽高
-  /// [downloadType] optional, 下载类广告是否展示二次确认弹窗
   const AndroidRewardedVideoConfig({
     required this.slotId,
     this.userId,
@@ -185,7 +175,6 @@ class AndroidRewardedVideoConfig implements Config {
     this.isSupportDeepLink = true,
     this.loadingType,
     this.expressSize,
-    this.downloadType = PangleDownloadType.none,
   });
 
   /// Convert config to json
@@ -204,7 +193,6 @@ class AndroidRewardedVideoConfig implements Config {
       'isSupportDeepLink': isSupportDeepLink,
       'loadingType': loadingType?.index,
       'expressSize': expressSize.toJson(),
-      'downloadType': downloadType.index,
     };
   }
 }
@@ -214,7 +202,6 @@ class AndroidBannerConfig implements Config {
   final bool isSupportDeepLink;
   final PangleExpressSize expressSize;
   final int? interval;
-  final PangleDownloadType downloadType;
 
   /// The feed ad config for Android
   ///
@@ -224,13 +211,11 @@ class AndroidBannerConfig implements Config {
   /// [interval] The carousel interval, in seconds, is set in the range of 30~120s,
   ///   and is passed during initialization. If it does not meet the requirements,
   ///   it will not be in carousel ad.
-  /// [downloadType] optional, 下载类广告是否展示二次确认弹窗
   const AndroidBannerConfig({
     required this.slotId,
     required this.expressSize,
     this.isSupportDeepLink = true,
     this.interval,
-    this.downloadType = PangleDownloadType.none,
   });
 
   /// Convert config to json
@@ -241,7 +226,6 @@ class AndroidBannerConfig implements Config {
       'isSupportDeepLink': isSupportDeepLink,
       'expressSize': expressSize.toJson(),
       'interval': interval,
-      'downloadType': downloadType.index,
     };
   }
 }
@@ -251,7 +235,6 @@ class AndroidNativeBannerConfig implements Config {
   final bool isSupportDeepLink;
   final PangleSize size;
   final int? interval;
-  final PangleDownloadType downloadType;
 
   /// The feed ad config for Android
   ///
@@ -261,13 +244,11 @@ class AndroidNativeBannerConfig implements Config {
   /// [interval] The carousel interval, in seconds, is set in the range of 30~120s,
   ///   and is passed during initialization. If it does not meet the requirements,
   ///   it will not be in carousel ad.
-  /// [downloadType] optional, 下载类广告是否展示二次确认弹窗
   const AndroidNativeBannerConfig({
     required this.slotId,
     required this.size,
     this.isSupportDeepLink = true,
     this.interval,
-    this.downloadType = PangleDownloadType.none,
   });
 
   /// Convert config to json
@@ -278,7 +259,6 @@ class AndroidNativeBannerConfig implements Config {
       'isSupportDeepLink': isSupportDeepLink,
       'size': size.toJson(),
       'interval': interval,
-      'downloadType': downloadType.index,
     };
   }
 }
@@ -288,7 +268,6 @@ class AndroidFeedConfig implements Config {
   final int? count;
   final bool isSupportDeepLink;
   final PangleExpressSize expressSize;
-  final PangleDownloadType downloadType;
 
   /// The feed ad config for Android
   ///
@@ -296,13 +275,11 @@ class AndroidFeedConfig implements Config {
   /// [count] It is recommended to request no more than 3 ads. The maximum is 10. default 3
   /// [isSupportDeepLink] optional. Whether to support deeplink.
   /// [expressSize] 期望信息流广告宽高
-  /// [downloadType] optional, 下载类广告是否展示二次确认弹窗
   const AndroidFeedConfig({
     required this.slotId,
     required this.expressSize,
     this.count,
     this.isSupportDeepLink = true,
-    this.downloadType = PangleDownloadType.none,
   });
 
   /// Convert config to json
@@ -313,7 +290,6 @@ class AndroidFeedConfig implements Config {
       'count': count,
       'isSupportDeepLink': isSupportDeepLink,
       'expressSize': expressSize.toJson(),
-      'downloadType': downloadType.index,
     };
   }
 }
@@ -322,19 +298,16 @@ class AndroidInterstitialConfig implements Config {
   final String slotId;
   final bool isSupportDeepLink;
   final PangleExpressSize expressSize;
-  final PangleDownloadType downloadType;
 
   /// The interstitial ad config for Android
   ///
   /// [slotId] required. The unique identifier of a interstitial ad.
   /// [isSupportDeepLink] optional. Whether to support deep link. default true.
   /// [expressSize] optional. 模板宽高
-  /// [downloadType] optional, 下载类广告是否展示二次确认弹窗
   const AndroidInterstitialConfig({
     required this.slotId,
     required this.expressSize,
     this.isSupportDeepLink = true,
-    this.downloadType = PangleDownloadType.none,
   });
 
   /// Convert config to json
@@ -344,7 +317,6 @@ class AndroidInterstitialConfig implements Config {
       'slotId': slotId,
       'isSupportDeepLink': isSupportDeepLink,
       'expressSize': expressSize.toJson(),
-      'downloadType': downloadType.index,
     };
   }
 }
@@ -355,7 +327,6 @@ class AndroidFullscreenVideoConfig implements Config {
   final PangleOrientation orientation;
   final PangleLoadingType loadingType;
   final PangleExpressSize? expressSize;
-  final PangleDownloadType downloadType;
 
   /// The full screen video ad config for Android
   ///
@@ -364,14 +335,12 @@ class AndroidFullscreenVideoConfig implements Config {
   /// [orientation] 设置期望视频播放的方向，默认[PangleOrientation.veritical]
   /// [loadingType] optional. 加载广告的类型，默认[PangleLoadingType.normal]
   /// [expressSize] optional. 模板宽高
-  /// [downloadType] optional, 下载类广告是否展示二次确认弹窗
   const AndroidFullscreenVideoConfig({
     required this.slotId,
     this.isSupportDeepLink = true,
     this.orientation = PangleOrientation.veritical,
     this.loadingType = PangleLoadingType.normal,
     this.expressSize,
-    this.downloadType = PangleDownloadType.none,
   });
 
   /// Convert config to json
@@ -387,7 +356,6 @@ class AndroidFullscreenVideoConfig implements Config {
       'orientation': orientation.index,
       'loadingType': loadingType.index,
       'expressSize': expressSize.toJson(),
-      'downloadType': downloadType.index,
     };
   }
 }