+++ pub.sh
@@ -11,11 +11,11 @@ case $PARAM in
 dev|develop)
   dart format .
   # flutter packages pub publish --dry-run
-  pub publish --dry-run
+  dart pub publish --dry-run
   ;;
 rel|release)
   # flutter packages pub publish --server=https://pub.dartlang.org
   export PUB_HOSTED_URL=https://pub.dev FLUTTER_STORAGE_BASE_URL=https://storage.pub.dev
-  pub publish
+  dart pub publish
   ;;
 esac